rename: book_detail 패키지명을 bookdetail로 변경
