+++ app/src/main/java/com/boostcamp/and03/data/datasource/remote/bookdetail/BookDetailRemoteDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.boostcamp.and03.data.datasource.remote.book_detail
+package com.boostcamp.and03.data.datasource.remote.bookdetail
 
 import com.boostcamp.and03.data.api.BookDetailApiService
 import com.boostcamp.and03.data.model.response.BookLookUpResponse
